Display retweets
In the initial version I skipped retweets as they need extra handling
(it's a tweet embedded within a tweet).

This adds some minimal handling, showing the original tweet author with
the retweeter (i.e. the person I follow) afterwards.

The tweet text is replaced by the retweet text as the tweet is prepended
with "RT @originalauthor".